requirements.txt
runpod~=1.7.9
transformers>=5.0.0
accelerate>=1.0.0
pillow>=10.0.0
sentencepiece>=0.2.0
safetensors>=0.4.0

